+++ Tutti/Tutorials/Models/DeferredTutorial.swift
@@ -1,0 +1,11 @@
+//
+//  DeferredTutorial.swift
+//  Tutti
+//
+//  Created by Daniel Saidi on 2018-06-19.
+//  Copyright © 2018 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public protocol DeferredTutorial: Tutorial, DeferredOnboarding { }