+++ Tutti/Tutorials/Models/StandardTutorial.swift
@@ -8,8 +8,29 @@
 
 /*
  
- This is a simple implementation of the `Tutorial` protocol.
- You can use it as is or as a base for any custom tutorials.
+ This is a simple `Tutorial` implementation. Use it as is or
+ as a base for any custom tutorials.
+ 
+ You can create a standard tutorial, either by providing all
+ property values, or by using a localization page indication
+ key and automatically calculate the page count depending on
+ the existing translations that exist.
+ 
+ When you use the location-based approach, this class starts
+ at page 0 and looks for the `pageIndicationKey` translation
+ in `Localized.strings` for that page. It will then bump the
+ index until it cannot find a translation for a certain page.
+ The default key indicator value is `title`, and the default
+ key segment separator is `_` but you can use any values you
+ like. If you use these values, however, a tutorial with the
+ identifier `myTutorial` will get two pages if the following
+ language keys exist in `Localized.strings`:
+ 
+ * tutorial_myTutorial_0_title
+ * tutorial_myTutorial_1_title
+ 
+ You can subclass this class and override the initializer if
+ you want this automatic handling to behave differently.
  
  */
 
@@ -31,6 +52,20 @@ open class StandardTutorial: Tutorial {
         self.userId = userId
         self.keySegmentSeparator = keySegmentSeparator
         self.persistence = persistence
+    }
+    
+    public init(
+        fromLocalizationWithPageIndicationKey key: String,
+        identifier: String,
+        userId: String? = nil,
+        keySegmentSeparator: String = "_",
+        persistence: OnboardingPersistence = UserDefaults.standard) {
+        self.identifier = identifier
+        self.pageCount = -1
+        self.userId = userId
+        self.keySegmentSeparator = keySegmentSeparator
+        self.persistence = persistence
+        pageCount = resolvePageCount(withPageIndicationKey: key)
     }
     
     
@@ -71,5 +106,19 @@ open class StandardTutorial: Tutorial {
         var segments = ["tutorial", identifier, "\(pageIndex)"]
         if key.count > 0 { segments.append(key) }
         return segments.joined(separator: keySegmentSeparator)
+    }
+}
+
+
+// MARK: - Private Functions
+
+fileprivate extension StandardTutorial {
+    
+    func resolvePageCount(withPageIndicationKey key: String) -> Int {
+        var index = 0
+        while translationExists(for: resourceName(for: key, at: index)) {
+            index += 1
+        }
+        return index
     }
 }